js: Promise-based AES-SIV API
Changes the `AESSIV` API to be asynchronous and promised-based.

This is futureproofing for a potential switch to WebCrypto. WebCrypto uses
asynchronous APIs everywhere (which, IMO, is a tad silly for cryptographic
primitives that finish in nanoseconds, but que sera sera).

+++ js/src/internal/polyfill/aes.ts
@@ -294,7 +294,7 @@ function rotw(w: number): number {
   return (w << 8) | (w >>> 24);
 }
 
-function expandKey(key: Uint8Array, encKey: Uint32Array, decKey?: Uint8Array): void {
+function expandKey(key: Uint8Array, encKey: Uint32Array, decKey?: Uint32Array): void {
   const nk = key.length / 4 | 0;
   const n = encKey.length;
   for (let i = 0; i < nk; i++) {
